+++ src/vm.cpp
@@ -1337,10 +1337,15 @@ bool VM::callPipeline(ARState &state, Value &&desc, const bool lastPipe, const F
 
   assert(pipeSize > 0);
 
-  PipeSet pipeSet;
+  PipeSet pipeSet; // for substitution/co-process
   PipeList pipes(pipeSize);
-  if (!pipeSet.openAll(forkKind) || !pipes.openAll()) {
-    raiseSystemError(state, errno, ERROR_PIPE);
+  PipeList childSetups(procSize); // for last-pipe
+  if (!pipeSet.openAll(forkKind) || !pipes.openAll() || (lastPipe && !childSetups.openAll())) {
+    const int errNum = errno;
+    pipeSet.closeAll();
+    pipes.closeAll();
+    childSetups.closeAll();
+    raiseSystemError(state, errNum, ERROR_PIPE);
     return false;
   }
 
@@ -1388,8 +1393,12 @@ bool VM::callPipeline(ARState &state, Value &&desc, const bool lastPipe, const F
         errNum = errno;
       }
     }
+    if (lastPipe && write(childSetups[procIndex][WRITE_PIPE], "1", 1) != 1) {
+      errNum = errno;
+    }
     pipeSet.closeAll();
     pipes.closeAll();
+    childSetups.closeAll();
 
     if (errNum) {
       raiseSystemError(state, errNum, ERROR_FD_SETUP);
@@ -1401,6 +1410,22 @@ bool VM::callPipeline(ARState &state, Value &&desc, const bool lastPipe, const F
     return true;
   } else if (procIndex == procSize) { // parent (last pipeline)
     if (lastPipe) {
+      /**
+       * wait for child process setup completion due to rance condition of `tcsetpgrp`
+       *
+       * ex. cmd0 & cmd1 | cmd2 | { fg; }
+       * Ideally `cmd1` will be foreground process and after that, `fg` will change `cmd0` with
+       * foreground. In some slow execution environment, `cmd1` has not be foreground before call
+       * `fg`. `fg` change `cmd0` with foreground, but `cmd1` will be foreground (race condition).
+       * As a result, `cmd0` still be background
+       */
+      for (unsigned int i = 0; i < procSize; i++) {
+        char b[1];
+        if (read(childSetups[i][READ_PIPE], b, 1) != 1) {
+        } // ignore error
+      }
+      childSetups.closeAll();
+
       /**
        * in last pipe, save current stdin before call dup2
        */
@@ -1438,6 +1463,9 @@ bool VM::callPipeline(ARState &state, Value &&desc, const bool lastPipe, const F
     for (unsigned int i = 0; i < procIndex; i++) {
       static_cast<void>(children[i].send(SIGKILL));
     }
+    pipeSet.closeAll();
+    pipes.closeAll();
+    childSetups.closeAll();
 
     raiseSystemError(state, EAGAIN, "fork failed");
     return false;