+++ test/interactive/interactive_test3.cpp
@@ -163,12 +163,12 @@ TEST_F(InteractiveTest, lastpipe_ctrlc3) {
   this->send(CTRL_Z);
   ASSERT_NO_FATAL_FAILURE(this->expect(promptAfterCtrlZ(PROMPT), "[1] + Stopped  sleep 3000\n"));
 
-  this->sendLine("var aa = sleep 1000 | sleep 2000 | { sleep 0.1; jobs; fg; }");
-  ASSERT_NO_FATAL_FAILURE(
-      this->expect(PROMPT + "var aa = sleep 1000 | sleep 2000 | { sleep 0.1; jobs; fg; }\n"
-                            "[1] + Stopped  sleep 3000\n"
-                            "[2]   Running  sleep 1000 | sleep 2000 | { sleep 0.1; jobs; fg; }\n"
-                            "sleep 3000\n"));
+  this->sendLine("var aa = sleep 1000 | sleep 2000 | { jobs; fg; }");
+  ASSERT_NO_FATAL_FAILURE(this->expect(PROMPT +
+                                       "var aa = sleep 1000 | sleep 2000 | { jobs; fg; }\n"
+                                       "[1] + Stopped  sleep 3000\n"
+                                       "[2]   Running  sleep 1000 | sleep 2000 | { jobs; fg; }\n"
+                                       "sleep 3000\n"));
 
   std::this_thread::sleep_for(std::chrono::milliseconds(500));
   this->send(CTRL_C);